requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
Werkzeug==3.0.1
python-dotenv==1.0.0
pandas==2.1.3
matplotlib==3.8.2
plotly==5.18.0
scikit-learn==1.3.2
